Add Ionic Angular setup and main application layout
This commit includes the initial setup for an Ionic Angular application named "campus". It comprises the main application layout using the Ionic framework with configurations for routings and UI components. Furthermore, the testing environment setup using Jasmine and Karma is included, along with adjustments to ESLint and TypeScript settings specific to the project.